refactor(nats): drop the object-store on-write SHA-256
The object store hashed every uploaded stream on the fly, but with file
bytes now encrypted before storage that hash covered ciphertext and no
caller read it. Callers that need a plaintext hash compute it themselves
upstream of encryption.

Remove HashingReader, the sha256 field/accessors on PutResult, and the
now-orphaned sha2/hex deps; put() just streams the reader through.
