+++ crates/nvisy-nats/src/object/mod.rs
@@ -23,7 +23,6 @@
 //! - [`PutResult`] - Result of upload operations with size and SHA-256 hash
 //! - [`GetResult`] - Result of download operations with streaming reader
 
-mod hashing_reader;
 mod object_bucket;
 mod object_data;
 mod object_key;
